APIMongo/app/requirements.txt
fastapi
pymongo[srv]
python-dotenv
uvicorn
pandas